+++ aikau/src/main/resources/alfresco/search/AlfSearchResult.js
@@ -133,6 +133,28 @@ define(["dojo/_base/declare",
       enableContextMenu: false,
 
       /**
+       * The prefix string that indicates the start of text to highlight.
+       * 
+       * @instance
+       * @type {string}
+       * @default
+       * @since 1.0.92
+       */
+      highlightPrefix: "\u0000",
+
+      /**
+       * This is the property within the [currentItem]{@link module:alfresco/core/CoreWidgetProcessing#highlightProperty}
+       * that should be used to identify the content with in the 
+       * [renderedValue]{@link module:alfresco/renderers/Property#renderedValue} to highlight.
+       * 
+       * @instance
+       * @type {string}
+       * @default 
+       * @since 1.0.92
+       */
+      highlightPostfix: "\u0003",
+
+      /**
        * This is the property within the [currentItem]{@link module:alfresco/core/CoreWidgetProcessing#highlightProperty}
        * that should be used to identify the content with.
        * 
@@ -140,6 +162,8 @@ define(["dojo/_base/declare",
        * @type {string}
        * @default 
        * @since 1.0.87
+       * @deprecated Since 1.0.92 - use [highlightPrefix]{@link module:alfresco/search/AlfSearchResult#highlightPrefix}
+       * and [highlightPostfix]{@link module:alfresco/search/AlfSearchResult#highlightPostfix}
        */
       highlightProperty: null,
 
@@ -153,6 +177,8 @@ define(["dojo/_base/declare",
        * @type {string}
        * @default 
        * @since 1.0.87
+       * @deprecated Since 1.0.92 - use [highlightPrefix]{@link module:alfresco/search/AlfSearchResult#highlightPrefix}
+       * and [highlightPostfix]{@link module:alfresco/search/AlfSearchResult#highlightPostfix}
        */
       highlightValue: null,
 
@@ -300,6 +326,7 @@ define(["dojo/_base/declare",
          this.createSiteRenderer();
          this.createPathRenderer();
          this.createSizeRenderer();
+         this.createContentSnippet();
          this.createWidgetsBelow();
          this.createActionsRenderer();
          this.createContextActionsWidget();
@@ -324,6 +351,34 @@ define(["dojo/_base/declare",
             allowedActions: (this.currentItem.type === "document" || this.currentItem.type === "folder") ? this.documentAndFolderActions : this.otherNodeActions,
             widgetsForActions: this.widgetsForActions
          }, this.actionsNode);
+      },
+
+      /**
+       * 
+       * @instance
+       * @since 1.0.92
+       */
+      createContentSnippet: function alfresco_search_AlfSearchResult__createContentSnippet() {
+         if (this.showSearchTermHighlights)
+         {
+            var content = lang.getObject("highlighting.content", false, this.currentItem);
+            if (content)
+            {
+               // jshint nonew:false
+               new Property({
+                  id: this.id + "_CONTENT_SNIPPET",
+                  currentItem: this.currentItem,
+                  pubSubScope: this.pubSubScope,
+                  propertyToRender: "highlighting.content",
+                  renderedValuePrefix: "\"",
+                  renderedValueSuffix: "...\"",
+                  highlightValue: this.highlightValue,
+                  highlightPrefix: this.showSearchTermHighlights ? this.highlightPrefix : null,
+                  highlightPostfix: this.showSearchTermHighlights ? this.highlightPostfix : null
+               }, this.contentNode);
+            }
+         }
+
       },
 
       /**
@@ -376,7 +431,9 @@ define(["dojo/_base/declare",
                pubSubScope: this.pubSubScope,
                propertyToRender: "description",
                renderSize: "small",
-               highlightValue: this.highlightValue
+               highlightValue: this.highlightValue,
+               highlightPrefix: this.showSearchTermHighlights ? this.highlightPrefix : null,
+               highlightPostfix: this.showSearchTermHighlights ? this.highlightPostfix : null
             }, this.descriptionNode);
          }
       },
@@ -431,7 +488,9 @@ define(["dojo/_base/declare",
             renderSize: "large",
             newTabOnMiddleOrCtrlClick: this.newTabOnMiddleOrCtrlClick,
             defaultNavigationTarget: this.navigationTarget,
-            highlightValue: this.highlightValue
+            highlightValue: this.highlightValue,
+            highlightPrefix: this.showSearchTermHighlights ? this.highlightPrefix : null,
+            highlightPostfix: this.showSearchTermHighlights ? this.highlightPostfix : null
          };
          if (this.navigationTarget)
          {
@@ -519,7 +578,9 @@ define(["dojo/_base/declare",
                },
                newTabOnMiddleOrCtrlClick: this.newTabOnMiddleOrCtrlClick,
                defaultNavigationTarget: this.navigationTarget,
-               highlightValue: this.highlightValue
+               highlightValue: this.highlightValue,
+               highlightPrefix: this.showSearchTermHighlights ? this.highlightPrefix : null,
+               highlightPostfix: this.showSearchTermHighlights ? this.highlightPostfix : null
             }, this.pathNode);
          }
       },
@@ -562,14 +623,16 @@ define(["dojo/_base/declare",
          var site = lang.getObject("site.title", false, this.currentItem);
          if (!site)
          {
-            domClass.add(this.siteRow, "hidden");
+            domClass.add(this.siteNode, "hidden");
+            domClass.add(this.siteSeparatorNode, "hidden");
          }
          else
          {
             // jshint nonew:false
             new PropertyLink({
                id: this.id + "_SITE",
-               renderedValueClass: "alfresco-renderers-Property pointer",
+               renderedValueClass: "alfresco-renderers-Property pointer alfresco-search-AlfSearchResult__site",
+               deemphasized: true,
                renderSize: "small",
                pubSubScope: this.pubSubScope,
                currentItem: this.currentItem,
@@ -587,7 +650,9 @@ define(["dojo/_base/declare",
                },
                newTabOnMiddleOrCtrlClick: this.newTabOnMiddleOrCtrlClick,
                defaultNavigationTarget: this.navigationTarget,
-               highlightValue: this.highlightValue
+               highlightValue: this.highlightValue,
+               highlightPrefix: this.showSearchTermHighlights ? this.highlightPrefix : null,
+               highlightPostfix: this.showSearchTermHighlights ? this.highlightPostfix : null
             }, this.siteNode);
          }
       },
@@ -602,7 +667,8 @@ define(["dojo/_base/declare",
          // We only show the size if it's not empty and at least one byte
          if (!this.currentItem.size || this.currentItem.size < 0)
          {
-            domClass.add(this.sizeRow, "hidden");
+            domClass.add(this.sizeNode, "hidden");
+            domClass.add(this.sizeSeparatorNode, "hidden");
          }
          else
          {
@@ -612,6 +678,7 @@ define(["dojo/_base/declare",
                currentItem : this.currentItem,
                pubSubScope : this.pubSubScope,
                label : this.message("faceted-search.doc-lib.value-prefix.size"),
+               deemphasized: true,
                renderSize: "small",
                sizeProperty : "size"
             }, this.sizeNode);
@@ -698,7 +765,9 @@ define(["dojo/_base/declare",
                renderSize: "small",
                renderedValuePrefix: "(",
                renderedValueSuffix: ")",
-               highlightValue: this.highlightValue
+               highlightValue: this.highlightValue,
+               highlightPrefix: this.showSearchTermHighlights ? this.highlightPrefix : null,
+               highlightPostfix: this.showSearchTermHighlights ? this.highlightPostfix : null
             }, this.titleNode);
          }
       },