+++ aikau/src/main/resources/alfresco/services/actions/CopyMoveService.js
@@ -62,6 +62,16 @@ define(["dojo/_base/declare",
       repoNodeRef: "alfresco://company/home",
 
       /**
+       * Indicates whether or not [CopyMoveService]{@link module:alfresco/services/CopyMoveService} should support the
+       * creation of links for files and folders
+       * @instance
+       * @type {boolean}
+       * @default
+       * @since 1.0.92
+       */
+      supportLinkCreation: false,
+
+      /**
        * URL to call to copy a document
        *
        * @instance
@@ -80,7 +90,18 @@ define(["dojo/_base/declare",
       moveAPI: "slingshot/doclib/action/move-to/node/",
 
       /**
-       * Sets up the service using the configuration provided. 
+       * URL to call to create a link to a document
+       *
+       * @instance
+       * @type {string}
+       * @default
+       * @since 1.0.92
+       */
+      createLinkAPI: "api/node/doclink/",
+
+
+      /**
+       * Sets up the service using the configuration provided.
        *
        * @instance
        * @since 1.0.32
@@ -98,7 +119,7 @@ define(["dojo/_base/declare",
        * @property [confirmButtonLabel] {String} - The label for the confirmation button
        * @property [singleItemMode] {Bool} - {@link alfresco/pickers/PickedItems:singleItemMode}
        */
-      
+
       /* This function handles the creation of dialogs for both copy and move actions because
        * they are identical apart from the config object (which can be optionally passed in).
        * It defaults to copy mode if missing.
@@ -107,6 +128,7 @@ define(["dojo/_base/declare",
        * @param {object} payload The action payload
        * @param {object} documents The documents selected for copy or move
        * @param {createCopyMoveDialogConfig} [config] The config object.
+       * @fires module:alfresco/core/topics#CREATE_DIALOG
        */
       createCopyMoveDialog: function alfresco_services_ActionService__createCopyMoveDialog(payload) {
          var documents = payload.documents || [];
@@ -124,11 +146,22 @@ define(["dojo/_base/declare",
                documents: documents,
                responseScope: payload.alfResponseScope
             };
+         var createLinkButtonLabel = payload.createLinkButtonLabel || "services.ActionService.createLink",
+             createLinkResponseTopic = this.generateUuid() + "_ALF_CREATE_LINK_LOCATION_PICKED",
+             createLinkPublishPayload = {
+                nodes: nodes,
+                documents: documents,
+                responseScope: payload.alfResponseScope
+             };
 
          var firstFileName = documents.length ? documents[0].fileName : null;
          var fileName = (nodes.length === 1 && firstFileName) ? firstFileName : this.message("services.ActionService.copyMoveTo.multipleFiles");
-         this._actionHandle = this.alfSubscribe(responseTopic, lang.hitch(this, this.onCopyMoveLocationSelected, urlPrefix, payload.copy), true);
-         this.alfPublish("ALF_CREATE_DIALOG_REQUEST", {
+         this._actionHandles = [
+             this.alfSubscribe(responseTopic, lang.hitch(this, this.onLocationSelected, urlPrefix, payload.copy, this.performAction), true),
+             this.alfSubscribe(createLinkResponseTopic, lang.hitch(this, this.onLocationSelected, urlPrefix, payload.copy, this.performCreateLinkAction), true)
+          ];
+
+         this.alfPublish(topics.CREATE_DIALOG, {
             dialogId: "ALF_COPY_MOVE_DIALOG",
             dialogTitle: this.message(dialogTitle, { 0: fileName}),
             handleOverflow: false,
@@ -138,11 +171,27 @@ define(["dojo/_base/declare",
                   config: {
                      singleItemMode: singleItemMode,
                      generatePubSubScope: true,
-                     repoNodeRef: this.repoNodeRef || "alfresco://company/home"
+                     repoNodeRef: this.repoNodeRef || "alfresco://company/home",
+                     supportLinkCreation : this.supportLinkCreation
                   }
                }
             ],
             widgetsButtons: [
+               {
+                  name: "alfresco/buttons/AlfButton",
+                  config: {
+                     label: createLinkButtonLabel,
+                     publishTopic: createLinkResponseTopic,
+                     publishPayload: createLinkPublishPayload,
+                     disableOnInvalidControls: true,
+                     validTopic: "ALF_PICKER_VALID",
+                     invalidTopic: "ALF_PICKER_INVALID",
+                     additionalCssClasses: "call-to-action",
+                     visibilityConfig: {
+                        initialValue: this.supportLinkCreation && payload.copy
+                     }
+                  }
+               },
                {
                   name: "alfresco/buttons/AlfButton",
                   config: {
@@ -167,17 +216,18 @@ define(["dojo/_base/declare",
       },
 
       /**
-       * Handles the actual copy or move XHR call
+       * Handles the actual copy,move or create link XHR call
        *
        * @instance
        * @param {string} urlPrefix The specific URL prefix to use for the action (e.g. either move or copy)
        * @param {boolean} copy A boolean indicating if this is a copy action or not
        * @param {object} payload The payload from confirmation of the action dialog
+       * @param {function} function to call: create link or copy/move
        */
-      onCopyMoveLocationSelected: function alfresco_services_ActionService__onCopyMoveLocationSelected(urlPrefix, copy, payload) {
-         this.alfUnsubscribeSaveHandles([this._actionCopyHandle]);
+      onLocationSelected: function alfresco_services_ActionService__onLocationSelected(urlPrefix, copy, actionToCall, payload) {
+         this.alfUnsubscribeSaveHandles(this._actionHandles);
 
-         // Get the locations to copy to and the documents to them...
+         // Get the locations to copy/move/create link to and the documents to them...
          if (payload.dialogContent)
          {
             when(payload.dialogContent, lang.hitch(this, function(content) {
@@ -191,18 +241,16 @@ define(["dojo/_base/declare",
                   }
                   else if (!documents || documents.length === 0)
                   {
-                     this.alfLog("error", "Documents to copy not specified.");
+                     this.alfLog("error", "Documents to copy/move not specified.");
                   }
                   else
                   {
                      var nodeRefs = NodeUtils.nodeRefArray(documents);
-                     array.forEach(locations, lang.hitch(this, this.performAction, nodeRefs, urlPrefix, copy, payload.alfResponseScope));
+                     array.forEach(locations, lang.hitch(this, actionToCall, nodeRefs, urlPrefix, copy, payload.alfResponseScope));
                   }
                }
             }));
          }
-
-         
       },
 
       /**
@@ -233,8 +281,36 @@ define(["dojo/_base/declare",
       },
 
       /**
+       * Makes an XHR call to create links for the selected documents to the location provided.
+       *
+       * @instance
+       * @param {array} nodeRefs An array of the nodes to create links for
+       * @param {string} urlPrefix The prefix to use in the action URL
+       * @param {boolean} copy A boolean indicating if this is a copy action or not
+       * @param {array} location  The location where links will be created
+       * @since 1.0.92
+       */
+      performCreateLinkAction: function alfresco_services_actions_CopyMoveService__performCreateLinkAction(nodeRefs, urlPrefix, copy, responseScope, location) {
+         var responseTopic = this.generateUuid();
+         var successSubscription = this.alfSubscribe(responseTopic + "_SUCCESS", lang.hitch(this, this.onCreateLinkActionSuccess), true);
+         var failureSubscription = this.alfSubscribe(responseTopic + "_FAILURE", lang.hitch(this, this.onCreateLinkActionFailure), true);
+         this.serviceXhr({
+            alfTopic: responseTopic,
+            subscriptionHandles: [successSubscription,failureSubscription],
+            responseScope: responseScope,
+            url: AlfConstants.PROXY_URI +  this.createLinkAPI + location.nodeRef.replace("://", "/"),
+            method: "POST",
+            data: {
+               destinationNodeRef: location.nodeRef,
+               multipleFiles: nodeRefs
+            }
+         });
+      },
+
+
+      /**
        * Handles successful actions in one of two ways. When an attempt is made to move or copy more than one node
-       * it is possible that only some of the nodes will be moved/copied successfuly. If all the nodes were moved or
+       * it is possible that only some of the nodes will be moved/copied successfully. If all the nodes were moved or
        * copied successfully then a simple notification is displayed indicating a successful action was completed, however
        * if a partial success was reported then a prompt is displayed listing the nodes that could not be copied or moved.
        *
@@ -263,7 +339,7 @@ define(["dojo/_base/declare",
             {
                failureMessage = payload.response.totalResults === 1 ? "copyMoveService.move.failure" : "copyMoveService.move.multiple.failure";
             }
-            
+
             this.alfServicePublish(topics.DISPLAY_PROMPT, {
                title: payload.requestConfig.copy ? this.message("copyMoveService.copy.failure.title") : this.message("copyMoveService.move.failure.title"),
                message: this.message(failureMessage)
@@ -299,6 +375,46 @@ define(["dojo/_base/declare",
       },
 
       /**
+       * Handles successful actions. When an attempt is made to create link to more than one node it is possible that
+       * only for some of the nodes it will be created successfully. If for all the nodes the link was created
+       * successfully then a simple notification is displayed indicating a successful action was completed, or partial
+       * completed.
+       *
+       * @instance
+       * @param {object} payload
+       * @fires module:alfresco/core/topics#DISPLAY_NOTIFICATION
+       * @fires module:alfresco/core/topics#DISPLAY_PROMPT
+       * @fires module:alfresco/core/topics#RELOAD_DATA_TOPIC
+       * @since 1.0.92
+       */
+      onCreateLinkActionSuccess: function alfresco_services_actions_CopyMoveService__onCreateLinkActionSuccess(payload) {
+         // jshint unused:false
+         var subscriptionHandles = lang.getObject("requestConfig.subscriptionHandles", false, payload);
+         if (subscriptionHandles)
+         {
+            this.alfUnsubscribeSaveHandles(subscriptionHandles);
+         }
+         if (payload.response.successCount === 0)
+         {
+            // See AKU-1068 - total failure returned as success...
+            var failureMessage;
+            failureMessage = payload.response.failureCount === 1 ? "copyMoveService.createLink.failure" : "copyMoveService.createLink.multiple.failure";
+
+            this.alfServicePublish(topics.DISPLAY_PROMPT, {
+               title: this.message("copyMoveService.createLink.failure.title"),
+               message: this.message(failureMessage)
+            });
+         }
+         else if (payload.response.overallSuccess === "true")
+         {
+            this.alfServicePublish(topics.DISPLAY_NOTIFICATION, {
+               message: this.message("copyMoveService.createLink.completeSuccess")
+            });
+         }
+         this.alfPublish(topics.RELOAD_DATA_TOPIC, {}, false, false, payload.requestConfig.responseScope);
+      },
+
+      /**
        * Handles failed actions by displaying a notification indicating that the action was not successful.
        *
        * @instance
@@ -316,6 +432,27 @@ define(["dojo/_base/declare",
          this.alfServicePublish(topics.DISPLAY_PROMPT, {
             title: payload.requestConfig.copy ? this.message("copyMoveService.copy.failure.title") : this.message("copyMoveService.move.failure.title"),
             message: payload.requestConfig.copy ? this.message("copyMoveService.copy.failure") : this.message("copyMoveService.move.failure")
+         });
+      },
+
+      /**
+       * Handles failed actions by displaying a notification indicating that the action was not successful.
+       *
+       * @instance
+       * @param {object} payload
+       * @fires module:alfresco/core/topics#DISPLAY_PROMPT
+       * @since 1.0.92
+       */
+      onCreateLinkActionFailure: function alfresco_services_actions_CopyMoveService__onCreateLinkActionFailure(payload) {
+         // jshint unused:false
+         var subscriptionHandles = lang.getObject("requestConfig.subscriptionHandles", false, payload);
+         if (subscriptionHandles)
+         {
+            this.alfUnsubscribeSaveHandles(subscriptionHandles);
+         }
+         this.alfServicePublish(topics.DISPLAY_PROMPT, {
+            title: this.message("copyMoveService.createLink.failure.title"),
+            message: this.message("copyMoveService.createLink.failure")
          });
       }
    });